RTIOD/project/requirements.txt
hydra-core
matplotlib
omegaconf
pillow
pycocotools
scikit-learn
scipy
tqdm
wandb
seaborn
huggingface_hub
ultralytics